big push. Assignments done till Automation. Project done till Maths

+++ Project/03.Project Temp/SDKProject/PhysicsDll.h
@@ -1,1 +1,1 @@
-extern "C" double WeightCalculator(double, int);
+extern "C" double WeightCalculator(double, double);
